+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for large-records
 
+## Unreleased
+
+* Removed library dependency on micro-lens
+* Rename module Data.Record.Lens.Micro -> Data.Record.Lens.VL
+
 ## 0.1.0.0 -- 2021-08-19
 
 * First public release